Added a method to properly set the owner ( at least the email ) to the schema map in elastic search Removed traces of creating a new column family. Map manager is the way to go. Added the get endpoint to the service resource that will allow us to do gets. Added a way for the endpoint to read the JSON response and did the specific call out so that we could work around having to make a service ( Which is bad code )
Returned an object with a terrible name to debug the response that gets returned.
Added AbstractCollectionService method that allows us to get the workaround to a normal call in order to get around path permission checking.
AbstractService has the workaround which is what needs to be done if you want to avoid using a schema.
Added a new service action for the work around.